Factor out subcomponents and ajax library
